misc changes
* spike grenades should display max count properly
* napalm should display max count properly
* reduced hb stun to 25% max rate and duration to 0.1
* strike rate for shotgun is buffed to 22.2% from 20% at lv. 10
* hard mode scales every 4 players instead of 2. one player now will still scale.
* when voting, who proposed the vote should now be clear. (address #49)
* when voting, after votes have been decided, they will properly reset.
* When a new map is loaded, the vote is reset. (Just in case.)
* bfgs rockets and grenades no longer removed after death

+++ src/combat/abilities/napalm.c
@@ -115,7 +115,7 @@ void napalm_remove (edict_t *self, qboolean print)
 
 		if (print)
 			safe_cprintf(self->owner, PRINT_HIGH, "%d/%d napalm grenades remaining.\n",
-				self->owner->num_napalm, NAPALM_MAX_COUNT);
+				self->owner->num_napalm, (int)NAPALM_MAX_COUNT);
 	}
 }
 
@@ -234,7 +234,7 @@ void Cmd_Napalm_f (edict_t *ent)
 	if (Q_strcasecmp (gi.args(), "count") == 0)
 	{
 		safe_cprintf(ent, PRINT_HIGH, "You have %d/%d napalm grenades.\n",
-			ent->num_napalm, NAPALM_MAX_COUNT);
+			ent->num_napalm, (int)NAPALM_MAX_COUNT);
 		return;
 	}
